refactor: extract SimilarIssueReviewService to its own file (#649 slice A)
First slice of #649 (ExportService.swift decomposition). Extracts the
SimilarIssueReviewService actor (315 lines) from ExportService.swift into a
new SimilarIssueReviewService.swift.

Byte-identical move; SHA-256 4f319ac62ed19e0a363bb0efec4b11e14562dbc55075430dad242980478e6bfb
verified against origin/main:ExportService.swift lines 352-666.

Reduces ExportService.swift from 666 lines to 350 lines.

The SimilarIssueReviewService is an independent actor with its own dedup
logic — no coupling to ExportService except shared model types (all in
BugNarrator module, no visibility changes needed).

Tests: 667 (unchanged).
